+++ src/backends/webgpu/package.json
@@ -32,7 +32,7 @@
     "yalc": "~1.0.0-pre.21"
   },
   "dependencies": {
-    "@tensorflow/tfjs-core": "1.0.4",
+    "@tensorflow/tfjs-core": "1.1.0",
     "@webgpu/shaderc": "0.0.4",
     "@webgpu/types": "0.0.3"
   }
