+++ src/backends/webgpu/src/index.ts
@@ -28,12 +28,12 @@ export const ready = (async () => {
   const adapter = await navigator.gpu.requestAdapter({});
   const device = await adapter.requestDevice({});
 
-  tf.ENV.registerBackend('webgpu', () => {
+  tf.registerBackend('webgpu', () => {
     return new WebGPUBackend(device, shaderc);
   }, 3 /*priority*/);
 
   // If registration succeeded, set the backend.
-  if (tf.ENV.findBackend('webgpu') != null) {
+  if (tf.findBackend('webgpu') != null) {
     tf.setBackend('webgpu');
   }
 })();
